+++ rec_api/recommend.py
@@ -4,6 +4,7 @@ import matplotlib.pyplot as plt
 import keras
 import os
 import csv
+import json
 
 class Recommender:
     def loadData(self, path, ids, names, nameIndex):
@@ -116,12 +117,15 @@ class Recommender:
         # Print formatting
         max_width = max([len(rindex[c]) for c in closest])
 
-        # Print the most similar and distances
-        # umesto da printa neka vrati listu imena, idjeva i metriku
-        for c in reversed(closest):
-            print('{index_name.capitalize()}: {rindex[c]:{max_width + 2}} Similarity: {dists[c]:.{2}}')
+        # Return the most similar and distances
+        results = []
 
-    def test(self):
+        for c in reversed(closest):
+            results.append([book_id_to_title[c].capitalize(), str(dists[c])])
+
+        return results
+
+    def test(self, titles):
         directory = os.path.dirname(os.path.realpath(__file__))
         model = self.loadModel(str(directory) + '/initial_training.h5')
         book_weights = self.initWeights(model)
@@ -131,7 +135,15 @@ class Recommender:
         name_to_tag_id = {}
         self.loadData(str(directory) + '/books.csv', book_id_to_title, title_to_book_id, 9)
         self.loadData(str(directory) + '/tags.csv', tag_id_to_name, name_to_tag_id, 1)
-        return self.find_similar('The Hunger Games', book_weights, title_to_book_id, book_id_to_title, name_to_tag_id, name_to_tag_id)
+
+        results = [];
+
+        for title in titles:
+            results.append(self.find_similar(title, book_weights, title_to_book_id, book_id_to_title, name_to_tag_id, name_to_tag_id))
+
+        return results
+
+        self.find_similar('The Hunger Games', book_weights, title_to_book_id, book_id_to_title, name_to_tag_id, name_to_tag_id)
         self.find_similar("Harry Potter and the Philosopher's Stone", book_weights, title_to_book_id, book_id_to_title, name_to_tag_id, name_to_tag_id, return_dist = True)
         self.find_similar('Война и миръ', book_weights, title_to_book_id, book_id_to_title, name_to_tag_id, name_to_tag_id, return_dist = True)
         return 'Jeeste'
