+++ rec_api/views.py
@@ -1,40 +1,25 @@
-from django.shortcuts import render
-from django.http import HttpResponse
 from django.http import JsonResponse
+from rest_framework import serializers, viewsets
+from rest_framework.decorators import api_view
 from rec_api.models import Books
-# import json
-# from django.conf import settings  
-# from .apps import  RecApiConfig
-# from rest_framework.decorators import api_view
-# import os
-# import time
-# import glob
-# import requests
-# from scipy import ndimage
-# from scipy.ndimage import zoom
-
-# Create your views here.
-def rec_api(request): #view function
-    return JsonResponse({'message': 'desi le pi'})
-
-def list_books():
-    books = Books.objects.all()
-
-    data = {
-        "books": books
-    }
-    return HttpResponse(list(data), mimetype='application/json')
+from rec_api.recommend import Recommender
 
 
-# @api_view(["POST"]) 
+# Serializers define the API representation.
+class BookSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Books
+        fields = ['id', 'title', 'author', 'year', 'publisher', 'imageurls', 'imageurlm', 'imageurll']
 
-# def check_result(request):
-#     #Get video file url
-#     url = request.POST.get('url') 
-#     head, tail = os.path.split(url)
+# ViewSets define the view behavior.
+class BookViewSet(viewsets.ModelViewSet):
+    queryset = Books.objects.all()
+    serializer_class = BookSerializer
 
-#     r = requests.get(url, allow_redirects=True)
-
-#     emo_jso = emo_disp.to_json(orient='values') 
-
-#     return JsonResponse(emo_jso, safe=False) 
+@api_view(['GET', 'POST', 'DELETE'])
+def recommend(request):
+    if request.method == 'GET':
+        recommender = Recommender();
+        books = request.query_params.getlist('books')
+        recommendations = recommender.test(books)
+        return JsonResponse(recommendations, safe=False)
